User must exist with admin role (#65)
* Add checks in controllers

* Add tests

* Update OpenAPI docs

* Move role deletion protection from 'delete' to 'detachFromUsers'

---------

+++ api/src/main/java/api/services/UserService.java
@@ -41,6 +41,8 @@ public class UserService implements UserDetailsService {
     @Autowired
     private UserMapper userMapper;
     @Autowired
+    private RoleService roleService;
+    @Autowired
     private S3ObjectService s3ObjectService;
     @Autowired
     private MimeTypeService mimeTypeService;
@@ -132,8 +134,7 @@ public class UserService implements UserDetailsService {
      */
     @Transactional(readOnly = true)
     public Set<User> getAllContainingRole(Role role) {
-        Set<User> user = userRepository.findAllByRolesContaining(role);
-        return user;
+        return userRepository.findAllByRolesContaining(role);
     }
 
     /**
@@ -144,9 +145,8 @@ public class UserService implements UserDetailsService {
      */
     @Transactional(readOnly = true)
     public User get(int id) {
-        User user = find(id)
+        return find(id)
             .orElseThrow(() -> EntityNotFoundException.fromUser(id));
-        return user;
     }
 
     /**
@@ -157,9 +157,8 @@ public class UserService implements UserDetailsService {
      */
     @Transactional(readOnly = true)
     public User get(String username) {
-        User user = find(username)
+        return find(username)
             .orElseThrow(() -> EntityNotFoundException.fromUser(username));
-        return user;
     }
 
     /**
@@ -206,8 +205,15 @@ public class UserService implements UserDetailsService {
      */
     @Transactional
     public User save(User user) {
-        if (!user.getRoles().stream().anyMatch(role -> role.getName().equals(RoleService.USER_ROLE_NAME))) {
+        if (!user.getRoles().contains(roleService.get(RoleService.USER_ROLE_NAME))) {
             throw new IllegalStateException("User must have user role.");
+        }
+        if (!user.getRoles().contains(roleService.get(RoleService.ADMIN_ROLE_NAME))
+            && getAllContainingRole(roleService.get(RoleService.ADMIN_ROLE_NAME)).size() == 1
+            && getAllContainingRole(roleService.get(RoleService.ADMIN_ROLE_NAME)).iterator().next().getId() 
+                == user.getId()
+        ) {
+            throw new IllegalStateException("User must exist in the system with admin role.");
         }
         User saved = userRepository.save(user);
         log.info("User saved: " + user);
@@ -224,6 +230,11 @@ public class UserService implements UserDetailsService {
         if (!exists(id)) {
             throw EntityNotFoundException.fromUser(id);
         }
+        if (getRoles(id).contains(roleService.get(RoleService.ADMIN_ROLE_NAME))
+            && getAllContainingRole(roleService.get(RoleService.ADMIN_ROLE_NAME)).size() == 1
+        ) {
+            throw new IllegalStateException("Last admin user cannot be removed from the system.");
+        }
         userRepository.deleteById(id);
         log.info("User deleted: " + id);
     }
@@ -235,6 +246,12 @@ public class UserService implements UserDetailsService {
      */
     @Transactional
     public void delete(User user) {
+        if (user.getRoles().contains(roleService.get(RoleService.ADMIN_ROLE_NAME))
+            && getAllContainingRole(roleService.get(RoleService.ADMIN_ROLE_NAME)).size() == 1
+        ) {
+            throw new IllegalStateException("Last admin user cannot be removed from the system.");
+
+        }
         userRepository.delete(user);
         log.info("User deleted: " + user);
     }
